Android: When opening an external stream, find the corresponding file in the database to find the last read position and load the saved bookmarks, and if not found, save it to the application's private directory and save the information to the database.

+++ android/src/org/coolreader/crengine/Utils.java
@@ -76,8 +76,8 @@ public class Utils {
 		return moveFile(oldPlace, newPlace, false);
 	}
 
-	public static int copyStreamContent(OutputStream os, InputStream is) throws IOException {
-		int totalSize = 0;
+	public static long copyStreamContent(OutputStream os, InputStream is) throws IOException {
+		long totalSize = 0;
 		byte[] buf = new byte[0x10000];
 		for (;;) {
 			int bytesRead = is.read(buf);
